Moved the remaining items
