+++ Jantimizer/DatabaseConnector/Connectors/DbConnector.cs
@@ -1,5 +1,4 @@
-﻿using DatabaseConnector.Models;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Common;
