+++ Jantimizer/DatabaseConnector/Connectors/MySqlConnector.cs
@@ -1,5 +1,4 @@
 ﻿using MySqlConnector;
-using DatabaseConnector.Models;
 using System.Data;
 using System.Text.RegularExpressions;
