+++ Jantimizer/DatabaseConnector/Connectors/PostgreSqlConnector.cs
@@ -1,5 +1,4 @@
 ﻿using Npgsql;
-using DatabaseConnector.Models;
 using System.Data;
 using System.Text.RegularExpressions;
