+++ Jantimizer/QueryPlanParser/IPlanParser.cs
@@ -1,5 +1,4 @@
-﻿using DatabaseConnector.Connectors;
-using DatabaseConnector.Models;
+﻿using QueryPlanParser.Models;
 using System;
 using System.Collections.Generic;
 using System.Data;
@@ -7,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DatabaseConnector.Parsers
+namespace QueryPlanParser
 {
     public interface IPlanParser
     {
