+++ Jantimizer/QueryPlanParser/Models/AnalysisResultWithIndent.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DatabaseConnector.Models
+namespace QueryPlanParser.Models
 {
     public class AnalysisResultWithIndent
     {
