+++ Jantimizer/QueryPlanParser/Parsers/BasePlanParser.cs
@@ -1,4 +1,4 @@
-﻿using DatabaseConnector.Models;
+﻿using QueryPlanParser.Models;
 using System;
 using System.Collections.Generic;
 using System.Data;
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DatabaseConnector.Parsers
+namespace QueryPlanParser.Parsers
 {
     public abstract class BasePlanParser : IPlanParser
     {
