+++ Jantimizer/QueryPlanParser/Parsers/PostgreSqlParser.cs
@@ -1,4 +1,4 @@
-﻿using DatabaseConnector.Models;
+﻿using QueryPlanParser.Models;
 using System;
 using System.Collections.Generic;
 using System.Data;
@@ -7,7 +7,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace DatabaseConnector.Parsers
+namespace QueryPlanParser.Parsers
 {
     public class PostgreSqlParser : BasePlanParser
     {
