+++ Jantimizer/QueryTestSuite/Models/DBConnectorParser.cs
@@ -1,20 +1,20 @@
-﻿using DatabaseConnector.Connectors;
-using DatabaseConnector.Parsers;
+﻿using DatabaseConnector;
+using QueryPlanParser;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DatabaseConnector
+namespace QueryTestSuite.Models
 {
-    public class DatabaseParserConnector : IDatabaseParserConnector
+    internal class DBConnectorParser
     {
         public string Name { get; set; }
         public IDbConnector Connector { get; set; }
         public IPlanParser Parser { get; set; }
 
-        public DatabaseParserConnector(string name, IDbConnector connector, IPlanParser parser)
+        public DBConnectorParser(string name, IDbConnector connector, IPlanParser parser)
         {
             Name = name;
             Connector = connector;
