+++ Jantimizer/QueryTestSuite/Program.cs
@@ -1,9 +1,9 @@
 ﻿using DatabaseConnector;
 using DatabaseConnector.Connectors;
-using DatabaseConnector.Models;
-using DatabaseConnector.Parsers;
 using QueryParser;
 using QueryParser.QueryParsers;
+using QueryPlanParser.Parsers;
+using QueryTestSuite.Models;
 using QueryTestSuite.Services;
 using QueryTestSuite.TestRunners;
 
@@ -20,16 +20,16 @@ namespace QueryTestSuite
         {
             SecretsService secrets = new SecretsService();
 
-            var postgresModel = new DatabaseParserConnector(
+            var postgresModel = new DBConnectorParser(
                 "postgre",
                 new PostgreSqlConnector(secrets.GetConnectionString("POSGRESQL")),
                 new PostgreSqlParser());
-            var mySqlModel = new DatabaseParserConnector(
+            var mySqlModel = new DBConnectorParser(
                 "mysql",
                 new DatabaseConnector.Connectors.MySqlConnector(secrets.GetConnectionString("MYSQL")),
                 new MySQLParser());
 
-            var connectorSet = new List<DatabaseParserConnector>() { postgresModel };
+            var connectorSet = new List<DBConnectorParser>() { postgresModel };
 
 
             string testBaseDirPath = Path.GetFullPath("../../../Tests");
